move cell data models to a different script - will add to cell defiitions and script will become long
