+++ tests/test_conflicts.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from py._path.local import LocalPath
 
 from databooks.conflicts import path2conflicts
-from databooks.data_models.notebook import Cell, CellMetadata, NotebookMetadata
+from databooks.data_models.cell import Cell, CellMetadata
+from databooks.data_models.notebook import NotebookMetadata
 from tests.test_data_models.test_notebook import TestJupyterNotebook
 from tests.test_git_utils import ConflictFile, init_repo_conflicts
